Added BaseConfig and PlayerConfig
Dockerfile Has the base configurations for the vnc player, RedDockerfile has wireshark and vim on it. The RedDockerfile builds off of the base configuration in order to save build time. These aren't pretty, but they won't be for a while. Right now, they just work.
Next goal is to make each set of tools and everything work for the separate roles and players.